**New** wordstrap/wordstrap directory
This was a result of making WordStrap a parent theme.

The way WP-LESS references the LESS files for compilation results in the wordstrap directory being created redundantly inside of the wordstrap theme directory.

This actually turns out to be a good thing because it separates out a lot of the WordStrap code from the Roots code. The WordStrap-specific assets reside in this new directory.

For example, the compiled CSS resulting from the LESS files in the "Roots"/WordStrap assets directory is published (and cached) in wordstrap/wordstrap/assets/css/less/.

The custom fonts (like Font Awesome) + the UI images referenced in the options panel and stylesheets are also in this directory.

Signed-off-by: Jason Loftis <jason@jloft.com>