lib/wp-less/lib/Stylesheet.class.php
This file needed editing after shifting WordStrap to being parent theme enabled.

The Roots Theme (which WordStrap is a fork of) adds all sorts of URL and Path Rewrites. This makes it a bit cumbersome to directly reference assets in the child theme. The fork by "leoj3n" adjusted the rewrite rules to stabilize parent/child theme compatibility.

But the Roots rewrite rules were conflicting with the WP-LESS rewrite rules when trying to use code like "get_stylesheet_directory_uri" to wp_enqueue_style a LESS file from the child theme.

The workaround required editing this file's "source_path" variable. The way I configured it, the theme directory must be specified in the URL strings provided in the "bootstrap-for-theme.php" file that compiles the LESS variables + files into the final cached CSS.

This resulted in redundant directories being created by WP-LESS upon output of the CSS files because it re-creates the original directory structure of the LESS files location when outputting the complied CSS.

Signed-off-by: Jason Loftis <jason@jloft.com>