lib/wp-less/bootstrap-for-theme.php
A few modifications were made to accommodate the transition of WordStrap to parent/child theme compatibility.

The "Roots" rewrite rules and the WP-LESS rewrite rules were clashing a bit, so the way the wp_enqueue_style code references the LESS files was altered to include explicit reference to the theme (or child theme) directory. (See the commit for the Stylesheet.class.php file.)

The upload directory settings for the WP-LESS output of the complied CSS had to be reconfigured to also explicitly reference the active theme directory.

As a result of this configuration, WordStrap may be activated directory as a starter theme or indirectly as a parent theme with a child theme activated.

Signed-off-by: Jason Loftis <jason@jloft.com>