EDIT admin/options.php
This is a complete overhaul of the options.php file.

Unique option types were created for selecting the background pattern for each major HTML area of the website template.

Employed a new function for generating the patterns arrays for the radio img option types.

Added the "Subtle Patterns" background patterns as an array of "opaque" backgrounds.

Added the "info toggle" and "close toggle" option types. These wrap an options section in the standard WP UI for an accordion/toggle style element. This greatly enhanced the organization of numerous option sections under a single heading tab.

Javascript files for the show/hide of options based on a selection were broken out into their own separate files and include_once references used to embed them.

Lots of detail enhancements to each option section for things like margins, box shadow, positioning, containment, etc.

Significant effort was given to reformatting the code to be easier to read and find items.

Signed-off-by: Jason Loftis <jason@jloft.com>