EDIT admin/css/options-custom.css
The edits are too numerous to list. Essentially, changes were made to keep up with all the interface alterations made to the options panel.

This file is now a complete override of the default options CSS file provided by the framework. I copied over the entire original file and edited from there.

Signed-off-by: Jason Loftis <jason@jloft.com>