EDIT lib/scripts.php
Now, the assets/js/plugins.js file is registered to be enqueued in the footer instead of the header.

Commented out the enqueue of the child theme stylesheet.

The child theme only needs its default style.css to register the theme with the WP-ADMIN. But it contains no styles that need to be included in the website templates.

All necessary child theme CSS is handled by its ws.app.less, ws.child-theme.less, and ws.framework.less files. These are processed by the wordstrap parent theme's bootstrap-for-theme.php.

Signed-off-by: Jason Loftis <jason@jloft.com>