UPDATE assets/css/less/ws.type.less
This is a custom iteration of the Bootstrap type.less file. This commit updates it to version 2.2.1.  WordStrap still chooses to do some things differently. It was good to see Bootstrap shift to determining the font size and line height for the headings as derivatives of the baseLineHeight and baseFontSize. But that was something WordStrap already shifted to doing in a slightly different manner. WordStrap implements the Page Header differently as well.

Signed-off-by: Jason Loftis <jason@jloft.com>